BUG: Fix partial content error in NID [skip ci]
Fix an issue in `NID.stage_nid_inventory` where the function was failing
when the response status code was 206 (partial content). This issue is fixed
by checking the response status code and if it's 206, the function will continue
reading the headers and the get the modified date from the response headers.